+++ test/functional/feature_pruning.py
@@ -86,17 +86,17 @@ class PruneTest(BitcoinTestFramework):
         # Create nodes 0 and 1 to mine.
         # Create node 2 to test pruning.
         self.full_node_default_args = ["-maxreceivebuffer=20000", "-blockmaxsize=999000",
-                                       "-checkblocks=5", "-noparkdeepreorg", "-maxreorgdepth=-1"]
+                                       "-checkblocks=5", "-noparkdeepreorg"]
         # Create nodes 3 and 4 to test manual pruning (they will be re-started with manual pruning later)
         # Create nodes 5 to test wallet in prune mode, but do not connect
         self.extra_args = [self.full_node_default_args,
                            self.full_node_default_args,
                            ["-maxreceivebuffer=20000", "-prune=550",
-                               "-noparkdeepreorg", "-maxreorgdepth=-1"],
+                               "-noparkdeepreorg"],
                            ["-maxreceivebuffer=20000", "-blockmaxsize=999000",
-                               "-noparkdeepreorg", "-maxreorgdepth=-1"],
+                               "-noparkdeepreorg"],
                            ["-maxreceivebuffer=20000", "-blockmaxsize=999000",
-                               "-noparkdeepreorg", "-maxreorgdepth=-1"],
+                               "-noparkdeepreorg"],
                            ["-prune=550"]]
         self.rpc_timeout = 120
 
@@ -383,7 +383,7 @@ class PruneTest(BitcoinTestFramework):
         # check that the pruning node's wallet is still in good shape
         self.log.info("Stop and start pruning node to trigger wallet rescan")
         self.restart_node(
-            2, extra_args=["-prune=550", "-noparkdeepreorg", "-maxreorgdepth=-1"])
+            2, extra_args=["-prune=550", "-noparkdeepreorg"])
         self.log.info("Success")
 
         # check that wallet loads successfully when restarting a pruned node after IBD.
@@ -393,7 +393,7 @@ class PruneTest(BitcoinTestFramework):
         nds = [self.nodes[0], self.nodes[5]]
         self.sync_blocks(nds, wait=5, timeout=300)
         self.restart_node(
-            5, extra_args=["-prune=550", "-noparkdeepreorg", "-maxreorgdepth=-1"])
+            5, extra_args=["-prune=550", "-noparkdeepreorg"])
         self.log.info("Success")
 
     def run_test(self):