+++ doc/release-notes.md
@@ -4,4 +4,6 @@ Bitcoin ABC version 0.26.12 is now available from:
 
   <https://download.bitcoinabc.org/0.26.12/>
 
-This is a maintenance release with no user-visible change.
+ - Removed 10-block rolling finalization. This only affects configurations
+   where Avalanche is explicitly disabled. This removal includes associated
+   command-line parameters `-finalizationdelay` and `-maxreorgdepth`.